+++ src/main/java/net/sf/maventaglib/checker/TagAttribute.java
@@ -24,6 +24,8 @@
 package net.sf.maventaglib.checker;
 
 import org.apache.commons.lang3.builder.CompareToBuilder;
+import org.apache.commons.lang3.builder.EqualsBuilder;
+import org.apache.commons.lang3.builder.HashCodeBuilder;
 import org.apache.commons.lang3.builder.ToStringBuilder;
 import org.apache.commons.lang3.builder.ToStringStyle;
 
@@ -116,6 +118,41 @@ public class TagAttribute
         return new CompareToBuilder().append( this.deprecated, object.deprecated ).append( this.name, object.name )
             .toComparison();
     }
+
+    /**
+     * @see java.lang.Object#equals(Object)
+     */
+    @Override
+    public boolean equals( Object obj )
+    {
+        if( obj == null )
+        {
+            return false;
+        }
+
+        if( obj == this )
+        {
+            return true;
+        }
+
+        if( obj.getClass() != getClass() )
+        {
+            return false;
+        }
+
+        TagAttribute rhs = (TagAttribute) obj;
+        return new EqualsBuilder().append( this.deprecated, rhs.deprecated ).append( this.name, rhs.name )
+                .isEquals();
+      }
+
+    /**
+     * @see java.lang.Object#hashCode()
+     */
+    @Override
+    public int hashCode() {
+        return new HashCodeBuilder().append( this.deprecated ).append( this.name )
+                .toHashCode();
+      }
 
     /**
      * @see java.lang.Object#toString()